Link posts and users table
linked users and posts table
there by accessed the posts conditionally, changes in the "/" route.

in case of create route, added a visibility box, which sets a post public if clicked.

and added the checkownership middleware for the edit, delete

although by not having edit and delete buttons for the non owner users we are restricting them performing such tasks yet checkownerhsip does take care of this on  a safer side

in index.ejs, edit and delete buttons are shown condtionally.